[ObjC] Split helpers_unittest in to a helpers and names test.
